Update patch set 1
Patch Set 1:

> ocp-4.4-vanilla example-cnf FAILURE https://www.distributed-ci.io/jobs/dff79037-0c63-4054-a58c-3a29bef7708d/jobStates
 > from https://softwarefactory-project.io/r/#/c/20840/

This failure is interesting, mirroring failed because of an error when trying to copy the container from the catalog list. In fact I have seen the same error with my other job, so it is not linked to your changes but it means mirroring can easily failed our builds!

Patch-set: 1