Missing d-o-a-a/plays/scripts/ entry in .spec file
For the usage of must-gather to retrieve logs from the OCP cluster, a new script was created in d-o-a-a/plays/scripts/get-logs in CR 22648.
However, that file is not being created when updating d-o-a-a to bring the latest changes.
This issue is because some config is missed in d-o-a-a.spec file. The lines included are similar to the ones included in d-o-a.spec file, as it also has get-logs script under plays/scripts path.

Change-Id: I143fdb4a1027c2e7e45d3490ee04bbd4bb8eb3a9