Reorder the teardown hook in case of error or failure
Under normal circumstances, teardown hook is the last one executed. But, in case of error or failure, it is executed before the error/failure process.
This is causing problems in, for example, the tnf_test_example, because the teardown just removes one of the declared namespaces. So, in case of error/failure, the logs cannot be gathered from that deleted namespaces.
Some reordering has been done as a result, also moving the fail tasks to the dci-openshift-app-agent.yml.
Moreover, error can also happen before creating the job, so a new task has been declared for handling this.

Change-Id: I878e8ad8c596631c6454bb7f46967f3e019f93d6