Review conditions to mirror images in disconnected environments
As there may be privaty registries that can be used also in connected environments, it is not true that, if provisionhost_registry and all_registries_creds are defined and length is higher than 0, we always have to mirror the images used in d-o-a-a.

This change also includes the check of dci_disconnected value to confirm we only mirror images when we are in disconnected environments, regardless of the value of the two variables aforementioned.

Change-Id: Ief1f349a80307f02a43cca813a2ae3e457fc6210