Add LOG_LEVEL variable to support tnf v3.x.x
I am just adding LOG_LEVEL to the tnf execution and maintaining TNF_LOG_LEVEL, so that both are created.
It is not a problem because they are both environment variables for tnf execution; depending on the version,
it will be used one or another.
Doing this in this way in order to easily remove it when removing backwards compatibility.

Change-Id: I6cfb6b910e1c9434dbbea1c8403a53bcc161af66