Remove requirements.yml
Dependencies (Packages/Roles/Modules) required are handled through RPM.
The recommended way to install the app-agent

Instructions in the README include ways to do a installation from the
repository without RPMs, dependencies/requirements are included in
there.

The requirements.yml is commonly used in ansible repositories to
provide the roles required[0], but in this case the RPMs and the
documentation provide the packages/roles/modules needed to run the
agent.

[0] https://docs.ansible.com/ansible/latest/galaxy/user_guide.html#installing-multiple-roles-from-a-file

Change-Id: I8871de4f8ea5f3c50ede8bfea04c8b1d461b50e8