fix default dci_gits_to_components to be a list
current jobs from d-o-a-a when not setting additional
directories from dci_config_dir are taking the default
which is not a list, and jobs are using the default
path /etc/dci-openshift-app-agent, split each letter
and use it as a single component. This patch includes
the default path in a list to avoid the issue.

Change-Id: Ib2b1db197dc042f702afdbfb1704354df47f4325