Enable tnf check commit SHA feature
This feature, disabled by default, check if the commit SHA in which the used tnf image is based on is the same than the tnf version we're providing as variable to the cnf-cert role to properly clone the cnf-certification-test repo. They should be the same, else there are problems in quay.io with tnf image. Good feature for daily testing.

Change-Id: I8a033b875112283d726e55fcf372be1e7aa9e4f1